Fix unwritable sniffle dwell time binding

+++ src/main/java/org/openpnp/machine/reference/ContactProbeNozzle.java
@@ -437,7 +437,7 @@ public class ContactProbeNozzle extends ReferenceNozzle {
         this.sniffleIncrementZ = sniffleIncrementZ;
     }
 
-    public long getSniffleDwellTime() {
+    public int getSniffleDwellTime() {
         return sniffleDwellTime;
     }
